DB Constructor
Created an SQL file for recreating the database. This must be done because the default value for the IsBeingLoaned column in the media table was acting wonky.

I think so long as you have the updated media.db file, you are good to go, but I'm including the SQL file anyway for good measure.